Massive reworking of most parts of LayerManager to make it asynchronous
The main thing that was done was removal of the `getNClosestLandmarks`
methods and instead replacing them with `updateToCenterLocation`
methods. Then there's no expectation of the caller to receive the
results and it's up to the LayerManager to send a notification when new
landmarks are loaded.

While I was in there, I cleaned up a few other things as well. GNLayers
no longer store information in a dictionary keyed on IDs, they're now
keyed simply on GNLandmarks.