Fixed memory management issue.
When calling `landmark.activeLayers = SOMETHING`, look at GNLandmark.
GNLandmark defines activeLayers as a @property with the "retain"
attribute. This means it calls `retain` on whatever object you set it
to be. Meaning we need to either pass it an autoreleased object or
release the object after we pass it in. This isn't the case, however
for things like `layerInfoByLandmarkID` within GNLayer. This isn't a
synthesized property, and therefor does no memory management for us, so
we need its retain count to remain 1 (as it is after alloc, init).

Take a look at: http://developer.apple.com/iphone/library/documentation/Cocoa/Conceptual/MemoryMgmt/MemoryMgmt.html